fix(benchmark): fix missing cast
parser.parse() takes a size_t, but size() is an int. Insert an explicit cast to
avoid compilation warnings.

+++ benchmark/benchmark-lsp/benchmark-config.cpp
@@ -389,7 +389,8 @@ Benchmark_Config Benchmark_Config::load() {
                 if (parser
                         .parse(reinterpret_cast<const char*>(
                                    package_json_content->data()),
-                               package_json_content->size())
+                               narrow_cast<std::size_t>(
+                                   package_json_content->size()))
                         .get(root) != ::simdjson::SUCCESS) {
                   std::fprintf(stderr, "error: %s: parsing JSON failed\n",
                                package_json_path);
